+++ src/xyz/oihana/schema/constants/traits/AppointmentsTrait.php
@@ -4,6 +4,7 @@ namespace xyz\oihana\schema\constants\traits;
 
 use xyz\oihana\schema\constants\traits\appointments\CustomerAppointmentTrait;
 use xyz\oihana\schema\constants\traits\appointments\FollowUpTrait;
+use xyz\oihana\schema\constants\traits\appointments\MeetingReportTrait;
 use xyz\oihana\schema\constants\traits\appointments\VisitReportTrait;
 
 /**
@@ -16,6 +17,7 @@ use xyz\oihana\schema\constants\traits\appointments\VisitReportTrait;
 trait AppointmentsTrait
 {
     use CustomerAppointmentTrait ,
-        FollowUpTrait    ,
-        VisitReportTrait ;
+        FollowUpTrait            ,
+        MeetingReportTrait       ,
+        VisitReportTrait         ;
 }