+++ src/xyz/oihana/schema/constants/traits/appointments/MeetingReportTrait.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace xyz\oihana\schema\constants\traits\appointments;
+
+/**
+ * The property name constants of the {@see \xyz\oihana\schema\appointments\MeetingReport} class.
+ *
+ * @package xyz\oihana\schema\constants\traits\appointments
+ * @author  Marc Alcaraz (eKameleon)
+ * @since   1.5.0
+ */
+trait MeetingReportTrait
+{
+    public const string TAGS   = 'tags'   ;
+    public const string TOPICS = 'topics' ;
+}